Item controller added,

+++ src/main/java/com/shop/repositary/ItemRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.shop.model.Item;
 
 @Repository
-public interface ItemRepo extends JpaRepository<Item, Integer>{
+public interface ItemRepo extends JpaRepository<Item, Long>{
 
 }
